updated conservation scores

+++ src/local_conservation_analysis_pipeline/s1setup_folder.py
@@ -14,9 +14,8 @@ def import_and_reindex_hits_df(hits_file):
     hits_df = pd.read_csv(hits_file)
     assert (
         "reference_index" not in hits_df.columns
-    ), "hits_df already has a reference_index column, I need to use that name for reindexing. Please rename the column."
+    ), "hits_df already has a reference_index column, I need to use that name for reindexing. Please rename the column or set `new_index`=False."
     hits_df = hits_df.reset_index(names="reference_index")
-    hits_df.to_csv(hits_file.replace(".csv", "_original_reindexed.csv"), index=False)
     return hits_df
 
 
@@ -53,6 +52,8 @@ def ortholog_analysis_setup(
     output_dict["orthogroups"] = {}
     for level in levels:
         output_dict["orthogroups"][level] = copy.deepcopy(database_files[level])
+        if "conservation_scores" not in output_dict["orthogroups"][level]:
+            output_dict["orthogroups"][level]["conservation_scores"] = {}
     output_dict["query_sequence"] = aln_2_query_seq(
         output_dict["orthogroups"][levels[0]]["alignment_file"],
         query_gene_id,
@@ -65,6 +66,8 @@ def ortholog_analysis_setup_search(
     database_key: dict,
 ):
     output_dict = ortholog_analysis_setup(table_row, database_key)
+    if "critical_error" in output_dict:
+        return output_dict
     # make sure that hit sequence is uppercase
     hit_sequence = table_row["hit_sequence"].upper()
     output_dict["hit_sequence"] = hit_sequence
@@ -78,6 +81,8 @@ def ortholog_analysis_setup_given_positions(
     database_key: dict,
 ):
     output_dict = ortholog_analysis_setup(table_row, database_key)
+    if "critical_error" in output_dict:
+        return output_dict
     hit_start_pos = int(table_row["hit start position"])
     hit_end_pos = int(table_row["hit end position"])
     output_dict["hit_start_position"] = hit_start_pos
@@ -108,7 +113,7 @@ def ortholog_analysis_setup_given_positions(
 #     return output_dicts
 
 
-def main(hits_file, database_key_file, output_folder, hit_search_method):
+def main(hits_file, database_key_file, output_folder, hit_search_method, new_index):
     output_folder = Path(output_folder)
     output_folder.mkdir(parents=True, exist_ok=True)
     if hit_search_method not in ["search", "given_positions"]:
@@ -116,7 +121,15 @@ def main(hits_file, database_key_file, output_folder, hit_search_method):
             f'hit_search_method must be "search" or "given_positions", not {hit_search_method}'
         )
 
-    hits_df = import_and_reindex_hits_df(hits_file)
+    if new_index:
+        hits_df = import_and_reindex_hits_df(hits_file)
+        hits_df.to_csv(output_folder / Path(hits_file).name.replace(".csv", "_original_reindexed.csv"), index=False)
+    else:
+        hits_df = pd.read_csv(hits_file)
+        assert (
+            "reference_index" in hits_df.columns
+        ), "hits_df should already have a reference_index column if new_index=False."
+
     with open(database_key_file, "r") as f:
         database_key = json.load(f)
 
@@ -132,6 +145,11 @@ def main(hits_file, database_key_file, output_folder, hit_search_method):
         output_file = (
             analysis_folder / f'{output_dict["reference_index"]}-{str(output_dict["query_gene_id"]).replace(":","_")}.json'
         )
+        if output_file.exists():
+            print(f"File already exists: {output_file}")
+            print("Skipping...")
+            print("clear the folder and rerun if you want to re-do the analysis")
+            continue
         with open(output_file, "w") as f:
             json.dump(output_dict, f, indent=4)
